+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 v0.5.4 (NEXT)
 -------------------
 * don't attempt running models when there are too few dose-groups
+* removed doses-dropped parameter from dataset init
+* Added new `drop_dose` method to Dataset
 
 v0.5.3 (2017-03-02)
 -------------------